fix install script

+++ merlinclash/install.sh
@@ -53,7 +53,6 @@ rm -rf /jffs/softcenter/bin/yq
 rm -rf /jffs/softcenter/*.yaml
 rm -rf /jffs/softcenter/webs/Module_merlinclash*
 rm -rf /jffs/softcenter/res/icon-merlinclash.png
-rm -rf /jffs/softcenter/res/clash-kcp.jpg
 rm -rf /jffs/softcenter/scripts/clash*
 
 find /jffs/softcenter/init.d/ -name "*clash.sh" | xargs rm -rf
@@ -112,16 +111,16 @@ rm -rf /tmp/clash* >/dev/null 2>&1
 
 echo_date clash插件安装成功！
 	#生成yamls.txt
-	cp -rf /jffs/softcenter/merlinclash/yaml_bak/*.yaml  /jffs/softcenter/merlinclash/
+	[ -f /jffs/softcenter/merlinclash/yaml_bak/*.yaml ] && cp -rf /jffs/softcenter/merlinclash/yaml_bak/*.yaml /jffs/softcenter/merlinclash/
 		
 	#生成新的txt文件
 
 	rm -rf /jffs/softcenter/merlinclash/yaml_bak/yamls.txt
 	echo_date "创建yaml文件列表"
 	#find $fp  -name "*.yaml" |sed 's#.*/##' >> $fp/yamls.txt
-	find /jffs/softcenter/merlinclash/yaml_bak  -name "*.yaml" |sed 's#.*/##' |sed '/^$/d' | awk -F'.' '{print $1}' >> /jffs/softcenter/merlinclash/yaml_bak/yamls.txt
+	[ -f /jffs/softcenter/merlinclash/yaml_bak/*.yaml ] && find /jffs/softcenter/merlinclash/yaml_bak  -name "*.yaml" |sed 's#.*/##' |sed '/^$/d' | awk -F'.' '{print $1}' >> /jffs/softcenter/merlinclash/yaml_bak/yamls.txt
 	#创建软链接
-	ln -s /jffs/softcenter/merlinclash/yaml_bak/yamls.txt /tmp/yamls.txt
+	[ -f /jffs/softcenter/merlinclash/yaml_bak/*.yaml ] && ln -s /jffs/softcenter/merlinclash/yaml_bak/yamls.txt /tmp/yamls.txt
 	#
 	echo_date "初始化配置文件处理完成"
 if [ "$merlinclash_enable" == "1" ];then
